Explicit cast.
git-svn-id: svn://svn.code.sf.net/p/crossfire/code/server/trunk@15574 282e977c-c81d-0410-88c4-b93c2d0d6712

+++ server/c_misc.c
@@ -2274,7 +2274,7 @@ void do_harvest(object *pl, int dir, object *skill) {
 
     /* Check the new item can fit into inventory.
      * Fixes bug #3060474: fishing puts more fishes into inventory than you can carry. */
-    if (pl->weight + pl->carrying + inv->weight > get_weight_limit(pl->stats.Str)) {
+    if (((uint32)(pl->weight + pl->carrying + inv->weight)) > get_weight_limit(pl->stats.Str)) {
         draw_ext_info_format(NDI_WHITE, 0, pl, MSG_TYPE_SKILL, MSG_TYPE_SKILL_FAILURE, "You are carrying too much to %s a %s.", skill->slaying, inv->name);
         return;
     }
